Email: Initial implementation
This sends an email to the admin whenever someone registers.

TODO: Modify registration procedure so that admin has to approve of the
account before it is created.
Alternatively, we could also create a system where users are given a
registration ID to create the account. The ID would be linked with their
name and email, so the system can verify the registrant. This
implementation would minimize admin intervention.

Signed-off-by: Tianrui Guo <tianruig@gmail.com>